Implement scan function
Also added the concept of invertedMarketData to simulate TradingView's execution
environment. `invertedMarketData` is an object that contains multiple series
where index 0 refers to the most recent candle and higher indices refer to past
candles.

Inverting the series this way makes it much easier to write functions that
operate on series, because all of their index 0s refer to the same time. There's
less need to coordinate indices between series of different lengths this way.