Write comments about possible implementation paths
I'm starting to think I don't have to do anything close to a proper order book.
For limit orders, I could mix the buys and sells into the same array, and just
scan through the whole thing when trying to execute orders.  The simulator won't
have a ton of orders to execute, so I don't have to organize things like a
proper exchange.  I should be able to get away with huge simplifications.