Make limit sell orders able to short
However, similar to market sell, I can't simultaneously close a position and
open the reverse position in a single order yet.  For now, I can live with that
restriction.  The strategies I write won't try to do that.

If I really needed to do something like that, I could get by with a workaround
that uses two limit orders that are adjacent to each other to close a position
and open a new position in the opposite direction.