Stub out a strategy uses divergence
My favorite trading setup that I've had the most success with is
to enter when I detect regular divergence on multiple timeframes
simultaneously.  Its hit rate seems high, and I'd like to verify
that through backtesting.  Also, when it misses, the losses can
be kept fairly small with surprisingly tight stop losses.

I want to combine this with good ways to switch between bullish and
bearish bias.  There are a couple of different ways I can think of
to do this, and they're all decent.  Fortunately, this doesn't have
to be super precise, becaues it's an entry criteria but just a way
to switch biases.  I intend to use these biases to increase and
decrease the strategies aggressiveness.