Add research strategies
These strategies are for doing market or indicator research and are not intended
to be used for real trading. Many of them won't even emit orders.  A lot of them
console.log their findings.

Initially, I was going to do this kind of work in a Debug strategy that I didn't
check in, but there's value in the research I want to do, and I don't want to
throw that code away. Thus, I made a research/ directory to be a home for these
kinds of strategies that help me learn about the market and my own strategy
execution system.  There's alot to explore here.