Fix a bug that made pipeline.aggregatorFn mess up
...on 1d and greater timeframe boundaries.
I wondered for days how my math could could be wrong,
but then I checked the 1h -> 4h aggregations and calculations,
and they were fine.  It was the 1d timeframe that was messing up
which made me think that it was a timezone issue on affecting the
1d boundary.  I had to fix issues on that boundary before, and
this was a spot that I didn't quite have right.