Add the market symbol to the orders before sending to the exchange
I prefer to let the Trader do it and leave this part out of the strategies.
I think strategies should be written in a market-neutral way.
They should not care what market they are trading, so it's
not the strategy's responsibility to put it in the order objects it emits.
The code that's using this should already know what it is and that code
can add it in itself.

I don't want strategies to be cluttered with more housekeeping duties than they
need.