Make indicators detect and use InvertedSeries
If the imd.close is using and InvertedSeries, the indicators use that as a
clue to use InvertedSeries for its own data.  It's slightly unfortunate that
indicators have to care about this detail.  To make them not care, I'd have
to expose key naming so that something outside of the indicators could setup
the InvertedSeries beforehand, but I didn't feel like doing that.

It's one of those situations where it might be more elegant to do it that way,
but it's a lot of work upfront.