Introduce variable position sizing via confidence
The allowedToLong and allowedToShort functions have returned booleans up until
now.  Going forward, instead of returning a boolean true, they may return a
number.  This number will be used as a multiplier for the default position size.
For example, if you're not confident with the long signal, you can return 0.5
which would make the state machine use half the position size it would normaly
use for a long.